refactor: 💡 move getfilteredchoices to generic command utils
